+++ core/database/migrations/2025_07_11_134113_create_modules_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('modules', function (Blueprint $table) {
             $table->id();
             $table->string('titre');
-            $table->integer('duree');
+            $table->float('duree');
             $table->string('categorie');
             $table->unsignedBigInteger('formation_id')->nullable();
             $table->foreign('formation_id')
